Cache things in memory instead
No need for file caching as long octane is used.
