+++ .env.dusk.local.example
@@ -11,5 +11,3 @@ DB_DATABASE_STORE=osu_store_test
 DB_DATABASE_UPDATES=osu_updates_test
 
 ES_INDEX_PREFIX=test_
-
-CACHE_DRIVER_LOCAL=array
