+++ .env.example
@@ -36,7 +36,6 @@ OSU_API_KEY=
 
 BROADCAST_DRIVER=log
 CACHE_DRIVER=file
-# CACHE_DRIVER_LOCAL=file
 SESSION_DRIVER=file
 # SESSION_DOMAIN=
 # SESSION_SECURE_COOKIE=false
