+++ app/Providers/AppServiceProvider.php
@@ -12,6 +12,7 @@ use App\Libraries\ChatFilters;
 use App\Libraries\CleanHTML;
 use App\Libraries\Groups;
 use App\Libraries\LayoutCache;
+use App\Libraries\LocalCacheManager;
 use App\Libraries\Mods;
 use App\Libraries\MorphMap;
 use App\Libraries\OsuAuthorize;
@@ -40,6 +41,7 @@ class AppServiceProvider extends ServiceProvider
         'clean-html' => CleanHTML::class,
         'groups' => Groups::class,
         'layout-cache' => LayoutCache::class,
+        'local-cache-manager' => LocalCacheManager::class,
         'mods' => Mods::class,
         'route-section' => RouteSection::class,
         'score-pins' => ScorePins::class,
@@ -60,8 +62,7 @@ class AppServiceProvider extends ServiceProvider
 
         Queue::after(function (JobProcessed $event) {
             app('OsuAuthorize')->resetCache();
-            app('groups')->incrementResetTicker();
-            app('chat-filters')->incrementResetTicker();
+            app('local-cache-manager')->incrementResetTicker();
 
             Datadog::increment(
                 config('datadog-helper.prefix_web').'.queue.run',
